+++ WS05/Utils.h
@@ -6,7 +6,7 @@
 
 namespace sdds {
    void extractToken(std::string& str, std::string& token, char delim);
-   void extractToken(std::string& str, unsigned int& token, char delim);
+   void extractToken(std::string& str, size_t& token, char delim);
    void extractToken(std::string& str, double& token, char delim);
    void trim(std::string& str);
